Move all methods related to GetFeaturesByBBox into the according QR
